pyproject.toml
[project]
name = "xrrna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "click>=8.1.8",
    "ncbi-acc-download>=0.2.9",
    "pandas>=2.2.3",
]
